Changed Color.COLORS from list to tuple
I was mainly figuring that a tuple requires slightly less overhead and better
represents the color selections as an immutable collection -- I would use a
frozenset but random.choice needs to be able to index into the collection and
neither set() nor frozenset() support indexing.

I also thought about turning Color.COLORS into a list dynamically when
random_color() is called, but that would incur a big time and memory overhead,
especially since random_color() seems to be called in inner loops. So that would
defeat the entire point of representing COLORS as a set.

Anyway, the point is this seems to be the best alternative outside of using a
non-builtin object, which I could do but I already think we're importing enough
cruft as it stands.